Write operations can request complete ignoring of all future subframes
The "discard" mechanism was intended to allow a write to a special
register to cause the device to not respond, and behave as if it had
not been written to.  However, the tunnel protocol effectively ignored
this, which allowed tunnel writes and reads to take effect, even if
the device didn't actually respond with the results.  Needless to say
this caused problems.